+++ bytey_byte_buffer/src/byte_buffer_write/arrays.rs
@@ -3,7 +3,7 @@ use crate::{byte_buffer::ByteBuffer, byte_buffer_write::ByteBufferWrite, error::
 impl<T: ByteBufferWrite, const N: usize> ByteBufferWrite for [T; N] {
     #[inline]
     fn write_to_buffer(&self, buffer: &mut ByteBuffer) -> Result<()> {
-        self.len().write_to_buffer(buffer)?;
+        (self.len() as u64).write_to_buffer(buffer)?;
 
         for e in self {
             e.write_to_buffer(buffer)?;
@@ -14,7 +14,7 @@ impl<T: ByteBufferWrite, const N: usize> ByteBufferWrite for [T; N] {
 
     #[inline]
     fn write_to_buffer_le(&self, buffer: &mut ByteBuffer) -> Result<()> {
-        self.len().write_to_buffer_le(buffer)?;
+        (self.len() as u64).write_to_buffer_le(buffer)?;
 
         for e in self {
             e.write_to_buffer_le(buffer)?;
@@ -25,7 +25,7 @@ impl<T: ByteBufferWrite, const N: usize> ByteBufferWrite for [T; N] {
 
     #[inline]
     fn write_to_buffer_be(&self, buffer: &mut ByteBuffer) -> Result<()> {
-        self.len().write_to_buffer_be(buffer)?;
+        (self.len() as u64).write_to_buffer_be(buffer)?;
 
         for e in self {
             e.write_to_buffer_be(buffer)?;
@@ -38,7 +38,7 @@ impl<T: ByteBufferWrite, const N: usize> ByteBufferWrite for [T; N] {
 impl<T: ByteBufferWrite> ByteBufferWrite for [T] {
     #[inline]
     fn write_to_buffer(&self, buffer: &mut ByteBuffer) -> Result<()> {
-        self.len().write_to_buffer(buffer)?;
+        (self.len() as u64).write_to_buffer(buffer)?;
 
         for e in self {
             e.write_to_buffer(buffer)?;
@@ -49,7 +49,7 @@ impl<T: ByteBufferWrite> ByteBufferWrite for [T] {
 
     #[inline]
     fn write_to_buffer_le(&self, buffer: &mut ByteBuffer) -> Result<()> {
-        self.len().write_to_buffer_le(buffer)?;
+        (self.len() as u64).write_to_buffer_le(buffer)?;
 
         for e in self {
             e.write_to_buffer_le(buffer)?;
@@ -60,7 +60,7 @@ impl<T: ByteBufferWrite> ByteBufferWrite for [T] {
 
     #[inline]
     fn write_to_buffer_be(&self, buffer: &mut ByteBuffer) -> Result<()> {
-        self.len().write_to_buffer_be(buffer)?;
+        (self.len() as u64).write_to_buffer_be(buffer)?;
 
         for e in self {
             e.write_to_buffer_be(buffer)?;
@@ -73,7 +73,7 @@ impl<T: ByteBufferWrite> ByteBufferWrite for [T] {
 impl<T: ByteBufferWrite> ByteBufferWrite for Vec<T> {
     #[inline]
     fn write_to_buffer(&self, buffer: &mut ByteBuffer) -> Result<()> {
-        self.len().write_to_buffer(buffer)?;
+        (self.len() as u64).write_to_buffer(buffer)?;
 
         for e in self {
             e.write_to_buffer(buffer)?;
@@ -84,7 +84,7 @@ impl<T: ByteBufferWrite> ByteBufferWrite for Vec<T> {
 
     #[inline]
     fn write_to_buffer_le(&self, buffer: &mut ByteBuffer) -> Result<()> {
-        self.len().write_to_buffer_le(buffer)?;
+        (self.len() as u64).write_to_buffer_le(buffer)?;
 
         for e in self {
             e.write_to_buffer_le(buffer)?;
@@ -95,7 +95,7 @@ impl<T: ByteBufferWrite> ByteBufferWrite for Vec<T> {
 
     #[inline]
     fn write_to_buffer_be(&self, buffer: &mut ByteBuffer) -> Result<()> {
-        self.len().write_to_buffer_be(buffer)?;
+        (self.len() as u64).write_to_buffer_be(buffer)?;
 
         for e in self {
             e.write_to_buffer_be(buffer)?;
@@ -108,7 +108,7 @@ impl<T: ByteBufferWrite> ByteBufferWrite for Vec<T> {
 impl<T: ByteBufferWrite> ByteBufferWrite for &Vec<T> {
     #[inline]
     fn write_to_buffer(&self, buffer: &mut ByteBuffer) -> Result<()> {
-        self.len().write_to_buffer(buffer)?;
+        (self.len() as u64).write_to_buffer(buffer)?;
 
         for e in *self {
             e.write_to_buffer(buffer)?;
@@ -119,7 +119,7 @@ impl<T: ByteBufferWrite> ByteBufferWrite for &Vec<T> {
 
     #[inline]
     fn write_to_buffer_le(&self, buffer: &mut ByteBuffer) -> Result<()> {
-        self.len().write_to_buffer_le(buffer)?;
+        (self.len() as u64).write_to_buffer_le(buffer)?;
 
         for e in *self {
             e.write_to_buffer_le(buffer)?;
@@ -130,7 +130,7 @@ impl<T: ByteBufferWrite> ByteBufferWrite for &Vec<T> {
 
     #[inline]
     fn write_to_buffer_be(&self, buffer: &mut ByteBuffer) -> Result<()> {
-        self.len().write_to_buffer_be(buffer)?;
+        (self.len() as u64).write_to_buffer_be(buffer)?;
 
         for e in *self {
             e.write_to_buffer_be(buffer)?;
